So I added AllegrexInstructions.java and Decoder.java using interface and generic stuff.
AllegrexInstructions is an interface where you find all the doXXX functions to implement for both Disassembler and Processor (interpreter).

Decoder is a generic class to instantiate in the class implementing AllegrexInstructions.

I plan to add a "Decoder< Processor > interpreter;" member in Processor class and call it in stepCpu(), this way :
  interpreter.process(this, Memory.get_instance().read32(pc));

Hope it would work.

git-svn-id: http://jpcsp.googlecode.com/svn/trunk@148 4ab978e6-3352-0410-996c-0b42db0051c8