Re-written the loader into new file Loader.java and Elf32.java. Removed loader code from Emulator.java and deleted FileManager.java.
The Loader now handles memory allocation for the space that the module consumes. Relevant changes have been made to ModuleMgrForUser to reflect this.

The Loader now stores a list of all loaded modules. Each time a module is loaded the list is iterated and an attempt is made at fixing unresolved imports. Currently the module info is stored in the jpcsp.ModuleContext class, it should be moved to the SceKernelModuleInfo class and also the list moved into jpcsp.HLE.kernel.managers.ModuleManager.

Replaced the ShFlags and ShType enums in Elf32SectionHeader with ints.


git-svn-id: http://jpcsp.googlecode.com/svn/trunk@633 4ab978e6-3352-0410-996c-0b42db0051c8