Simple but massive refactoring. Use stringbuilder instead of stringbuffer in buffers that do not "escape" to class variables.
There is a stringBuffer variable that escapes (it is a field) but i'm not sure if it can be replaced - the function that modifies it and returns its toString() is never called. It is in jpcsp.log.HTMLLayout. I think the plan for this class might be to call it in another thread than the EDT. As such i will not modifiy the field.

git-svn-id: http://jpcsp.googlecode.com/svn/trunk@1309 4ab978e6-3352-0410-996c-0b42db0051c8